refactor: split build logic convention plugins
